perf: max_chunk 50k + softplus bf16 + conditional div_loss
3 performance optimizations:

1. max_chunk_size 8192 -> 50000: know N_local=50000 now runs as 1
   chunk instead of 4, eliminating scan/dynamic_slice overhead.
   HBM headroom is 18.58GB so single-chunk fits easily.

2. softplus bf16: dead neuron path in gate_srw_step now computes
   softplus directly in bf16 instead of cast->f32->softplus->bf16.
   At 1e-4 scale on negative values, bf16 precision is sufficient.

3. Conditional diversity loss: compute_spatial_diversity_loss runs
   only every 100 steps via jax.lax.cond. O(N^2) cosine similarity
   over 100k neurons is expensive; skipping 99% of steps saves ~3-5%.
   train_step now takes a `step` argument for this.

+++ models/dawn_spatial_v3.py
@@ -181,8 +181,7 @@ def make_sharded_srw(mesh, max_chunk_size=2048):
             sc = h_bf @ ec.T
             raw = sc - tau_bf
             gc = jnp.where(raw > 0, raw,
-                            (1e-4 * jax.nn.softplus(
-                                raw.astype(jnp.float32))).astype(jnp.bfloat16))
+                            1e-4 * jax.nn.softplus(raw))
             gc = jnp.clip(gc, 0.0, 10.0)
             eg = (jnp.exp(gc.astype(jnp.float32)) - 1.0).astype(jnp.bfloat16)
             ef = eg.astype(jnp.float32)
@@ -282,8 +281,7 @@ def make_sharded_srw_paired(mesh, max_chunk_size=2048):
             sc = jnp.einsum('bsrd,nd->bsrn', h_bf, ec)
             raw = sc - tau_bf  # [B,S,2,cs]
             gc = jnp.where(raw > 0, raw,
-                            (1e-4 * jax.nn.softplus(
-                                raw.astype(jnp.float32))).astype(jnp.bfloat16))
+                            1e-4 * jax.nn.softplus(raw))
             gc = jnp.clip(gc, 0.0, 10.0)
             eg = (jnp.exp(gc.astype(jnp.float32)) - 1.0).astype(jnp.bfloat16)  # [B,S,2,cs]
             ef = eg.astype(jnp.float32)
@@ -359,7 +357,7 @@ def _srw_chunked(x, h, emb_norm, tau_offset, w_read, w_write, n_chunks):
         sc = h_bf @ ec.T
         raw = sc - tau_bf
         gc = jnp.where(raw > 0, raw,
-                        (1e-4*jax.nn.softplus(raw.astype(jnp.float32))).astype(jnp.bfloat16))
+                        1e-4 * jax.nn.softplus(raw))
         gc = jnp.clip(gc, 0.0, 10.0)
         eg = (jnp.exp(gc.astype(jnp.float32))-1.0).astype(jnp.bfloat16)
         ef = eg.astype(jnp.float32)
